hooked up space destruction procedures for groups and subGroups
